requirements.txt
requests_toolbelt
configparser
PySide2
jinja2
